Update MilkyStream_dict.py
The first part of the code is only to read the data (from http://cdn.gea.esac.esa.int/Gaia/gdr2/gaia_source_with_rv/csv/ ) and transform them into cartesian coordianes. Because this takes a lot of time these results are saved and can be loaded again when they are needed. This saves a lot of time